+++ main.go
@@ -20,6 +20,8 @@ import (
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/serializer"
 	"k8s.io/client-go/rest"
 	restclient "k8s.io/client-go/rest"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
@@ -131,6 +133,10 @@ func mainError() error {
 			return microerror.Mask(errors.New("couldn't wait for cache sync"))
 		}
 	}
+
+	scheme := runtime.NewScheme()
+	codecs := serializer.NewCodecFactory(scheme)
+	universalDeserializer := codecs.UniversalDeserializer()
 
 	var resourceSkusClient compute.ResourceSkusClient
 	{
@@ -338,52 +344,16 @@ func mainError() error {
 		}
 	}
 
-	var clusterCreateMutator *cluster.CreateMutator
+	var clusterWebhookHandler *cluster.WebhookHandler
 	{
-		conf := cluster.CreateMutatorConfig{
+		c := cluster.WebhookHandlerConfig{
 			BaseDomain: cfg.BaseDomain,
 			CtrlClient: ctrlClient,
+			CtrlReader: ctrlCache,
+			Decoder:    universalDeserializer,
 			Logger:     newLogger,
 		}
-		clusterCreateMutator, err = cluster.NewCreateMutator(conf)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-	}
-
-	var clusterUpdateMutator *cluster.UpdateMutator
-	{
-		conf := cluster.UpdateMutatorConfig{
-			CtrlCache:  ctrlCache,
-			CtrlClient: ctrlClient,
-			Logger:     newLogger,
-		}
-		clusterUpdateMutator, err = cluster.NewUpdateMutator(conf)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-	}
-
-	var clusterCreateValidator *cluster.CreateValidator
-	{
-		c := cluster.CreateValidatorConfig{
-			BaseDomain: cfg.BaseDomain,
-			CtrlClient: ctrlClient,
-			Logger:     newLogger,
-		}
-		clusterCreateValidator, err = cluster.NewCreateValidator(c)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-	}
-
-	var clusterUpdateValidator *cluster.UpdateValidator
-	{
-		c := cluster.UpdateValidatorConfig{
-			CtrlClient: ctrlClient,
-			Logger:     newLogger,
-		}
-		clusterUpdateValidator, err = cluster.NewUpdateValidator(c)
+		clusterWebhookHandler, err = cluster.NewWebhookHandler(c)
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -448,6 +418,30 @@ func mainError() error {
 		}
 	}
 
+	var validatorHttpHandlerFactory *validator.HttpHandlerFactory
+	{
+		c := validator.HttpHandlerFactoryConfig{
+			CtrlClient: ctrlClient,
+			CtrlCache:  ctrlCache,
+		}
+		validatorHttpHandlerFactory, err = validator.NewHttpHandlerFactory(c)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	var mutatorHttpHandlerFactory *mutator.HttpHandlerFactory
+	{
+		c := mutator.HttpHandlerFactoryConfig{
+			CtrlClient: ctrlClient,
+			CtrlCache:  ctrlCache,
+		}
+		mutatorHttpHandlerFactory, err = mutator.NewHttpHandlerFactory(c)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
 	// Here we register our endpoints.
 	handler := http.NewServeMux()
 	// Mutators.
@@ -457,8 +451,8 @@ func mainError() error {
 	handler.Handle("/mutate/azuremachinepool/update", mutator.Handler(azureMachinePoolUpdateMutator))
 	handler.Handle("/mutate/azurecluster/create", mutator.Handler(azureClusterCreateMutator))
 	handler.Handle("/mutate/azurecluster/update", mutator.Handler(azureClusterUpdateMutator))
-	handler.Handle("/mutate/cluster/create", mutator.Handler(clusterCreateMutator))
-	handler.Handle("/mutate/cluster/update", mutator.Handler(clusterUpdateMutator))
+	handler.Handle("/mutate/cluster/create", mutatorHttpHandlerFactory.NewCreateHandler(clusterWebhookHandler))
+	handler.Handle("/mutate/cluster/update", mutatorHttpHandlerFactory.NewUpdateHandler(clusterWebhookHandler))
 	handler.Handle("/mutate/machinepool/create", mutator.Handler(machinePoolCreateMutator))
 	handler.Handle("/mutate/machinepool/update", mutator.Handler(machinePoolUpdateMutator))
 	handler.Handle("/mutate/spark/create", mutator.Handler(sparkCreateMutator))
@@ -472,8 +466,8 @@ func mainError() error {
 	handler.Handle("/validate/azuremachine/update", validator.Handler(azureMachineUpdateValidator))
 	handler.Handle("/validate/azuremachinepool/create", validator.Handler(azureMachinePoolCreateValidator))
 	handler.Handle("/validate/azuremachinepool/update", validator.Handler(azureMachinePoolUpdateValidator))
-	handler.Handle("/validate/cluster/create", validator.Handler(clusterCreateValidator))
-	handler.Handle("/validate/cluster/update", validator.Handler(clusterUpdateValidator))
+	handler.Handle("/validate/cluster/create", validatorHttpHandlerFactory.NewCreateHandler(clusterWebhookHandler))
+	handler.Handle("/validate/cluster/update", validatorHttpHandlerFactory.NewUpdateHandler(clusterWebhookHandler))
 	handler.Handle("/validate/machinepool/create", validator.Handler(machinePoolCreateValidator))
 	handler.Handle("/validate/machinepool/update", validator.Handler(machinePoolUpdateValidator))
 	handler.HandleFunc("/healthz", healthCheck)