+++ pkg/cluster/validate_update.go
@@ -6,67 +6,30 @@ import (
 
 	aeconditions "github.com/giantswarm/apiextensions/v3/pkg/conditions"
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
 	capi "sigs.k8s.io/cluster-api/api/v1alpha3"
 	capiconditions "sigs.k8s.io/cluster-api/util/conditions"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/conditions"
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
 	"github.com/giantswarm/azure-admission-controller/internal/releaseversion"
 	"github.com/giantswarm/azure-admission-controller/internal/semverhelper"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/validator"
+	"github.com/giantswarm/azure-admission-controller/pkg/key"
 )
 
-type UpdateValidator struct {
-	ctrlClient client.Client
-	logger     micrologger.Logger
-}
-
-type UpdateValidatorConfig struct {
-	CtrlClient client.Client
-	Logger     micrologger.Logger
-}
-
-func NewUpdateValidator(config UpdateValidatorConfig) (*UpdateValidator, error) {
-	if config.CtrlClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
-	}
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-
-	v := &UpdateValidator{
-		ctrlClient: config.CtrlClient,
-		logger:     config.Logger,
-	}
-
-	return v, nil
-}
-
-func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.AdmissionRequest) error {
-	clusterNewCR := &capi.Cluster{}
-	clusterOldCR := &capi.Cluster{}
-	if _, _, err := validator.Deserializer.Decode(request.Object.Raw, nil, clusterNewCR); err != nil {
-		return microerror.Maskf(errors.ParsingFailedError, "unable to parse Cluster CR: %v", err)
-	}
-	if _, _, err := validator.Deserializer.Decode(request.OldObject.Raw, nil, clusterOldCR); err != nil {
-		return microerror.Maskf(errors.ParsingFailedError, "unable to parse Cluster CR: %v", err)
-	}
-
-	if !clusterNewCR.GetDeletionTimestamp().IsZero() {
-		a.logger.LogCtx(ctx, "level", "debug", "message", "The object is being deleted so we don't validate it")
-		return nil
-	}
-
-	capi, err := generic.IsCAPIRelease(clusterNewCR)
+func (h *WebhookHandler) OnUpdateValidate(ctx context.Context, oldObject interface{}, object interface{}) error {
+	clusterNewCR, err := key.ToClusterPtr(object)
 	if err != nil {
 		return microerror.Mask(err)
 	}
-	if capi {
+	if !clusterNewCR.GetDeletionTimestamp().IsZero() {
+		h.logger.LogCtx(ctx, "level", "debug", "message", "The object is being deleted so we don't validate it")
 		return nil
+	}
+
+	clusterOldCR, err := key.ToClusterPtr(oldObject)
+	if err != nil {
+		return microerror.Mask(err)
 	}
 
 	err = clusterNewCR.ValidateUpdate(clusterOldCR)
@@ -94,11 +57,7 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Mask(err)
 	}
 
-	return a.validateRelease(ctx, clusterOldCR, clusterNewCR)
-}
-
-func (a *UpdateValidator) Log(keyVals ...interface{}) {
-	a.logger.Log(keyVals...)
+	return h.validateRelease(ctx, clusterOldCR, clusterNewCR)
 }
 
 func validateClusterNetworkUnchanged(old capi.Cluster, new capi.Cluster) error {
@@ -138,7 +97,7 @@ func validateClusterNetworkUnchanged(old capi.Cluster, new capi.Cluster) error {
 	return nil
 }
 
-func (a *UpdateValidator) validateRelease(ctx context.Context, clusterOldCR *capi.Cluster, clusterNewCR *capi.Cluster) error {
+func (h *WebhookHandler) validateRelease(ctx context.Context, clusterOldCR *capi.Cluster, clusterNewCR *capi.Cluster) error {
 	oldClusterVersion, err := semverhelper.GetSemverFromLabels(clusterOldCR.Labels)
 	if err != nil {
 		return microerror.Maskf(errors.ParsingFailedError, "unable to parse version from the Cluster being updated")
@@ -157,5 +116,5 @@ func (a *UpdateValidator) validateRelease(ctx context.Context, clusterOldCR *cap
 		}
 	}
 
-	return releaseversion.Validate(ctx, a.ctrlClient, oldClusterVersion, newClusterVersion)
+	return releaseversion.Validate(ctx, h.ctrlClient, oldClusterVersion, newClusterVersion)
 }