+++ pkg/cluster/validate_update_test.go
@@ -7,19 +7,17 @@ import (
 	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	capi "sigs.k8s.io/cluster-api/api/v1alpha3"
 
 	builder "github.com/giantswarm/azure-admission-controller/internal/test/cluster"
+	"github.com/giantswarm/azure-admission-controller/pkg/unittest"
 )
 
 func TestClusterUpdateValidate(t *testing.T) {
 	type testCase struct {
 		name         string
-		oldCluster   []byte
-		newCluster   []byte
+		oldCluster   *capi.Cluster
+		newCluster   *capi.Cluster
 		errorMatcher func(err error) bool
 	}
 
@@ -36,32 +34,32 @@ func TestClusterUpdateValidate(t *testing.T) {
 	var testCases = []testCase{
 		{
 			name:         "case 0: unchanged ControlPlaneEndpoint",
-			oldCluster:   clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
-			newCluster:   clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			oldCluster:   clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			newCluster:   clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
 			errorMatcher: nil,
 		},
 		{
 			name:         "case 1: host changed",
-			oldCluster:   clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
-			newCluster:   clusterRawObject("ab123", clusterNetwork, "api.azure.gigantic.io", 443, nil),
+			oldCluster:   clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			newCluster:   clusterObject("ab123", clusterNetwork, "api.azure.gigantic.io", 443, nil),
 			errorMatcher: IsControlPlaneEndpointWasChangedError,
 		},
 		{
 			name:         "case 2: port changed",
-			oldCluster:   clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
-			newCluster:   clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 80, nil),
+			oldCluster:   clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			newCluster:   clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 80, nil),
 			errorMatcher: IsControlPlaneEndpointWasChangedError,
 		},
 		{
 			name:         "case 3: clusterNetwork deleted",
-			oldCluster:   clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
-			newCluster:   clusterRawObject("ab123", nil, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			oldCluster:   clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			newCluster:   clusterObject("ab123", nil, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
 			errorMatcher: IsClusterNetworkWasChangedError,
 		},
 		{
 			name:       "case 4: clusterNetwork.APIServerPort changed",
-			oldCluster: clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
-			newCluster: clusterRawObject(
+			oldCluster: clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			newCluster: clusterObject(
 				"ab123",
 				&capi.ClusterNetwork{
 					APIServerPort: to.Int32Ptr(80),
@@ -80,8 +78,8 @@ func TestClusterUpdateValidate(t *testing.T) {
 		},
 		{
 			name:       "case 5: clusterNetwork.ServiceDomain changed",
-			oldCluster: clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
-			newCluster: clusterRawObject(
+			oldCluster: clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			newCluster: clusterObject(
 				"ab123",
 				&capi.ClusterNetwork{
 					APIServerPort: to.Int32Ptr(443),
@@ -100,8 +98,8 @@ func TestClusterUpdateValidate(t *testing.T) {
 		},
 		{
 			name:       "case 6: clusterNetwork.Services deleted",
-			oldCluster: clusterRawObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
-			newCluster: clusterRawObject(
+			oldCluster: clusterObject("ab123", clusterNetwork, "api.ab123.test.westeurope.azure.gigantic.io", 443, nil),
+			newCluster: clusterObject(
 				"ab123",
 				&capi.ClusterNetwork{
 					APIServerPort: to.Int32Ptr(443),
@@ -116,8 +114,8 @@ func TestClusterUpdateValidate(t *testing.T) {
 		},
 		{
 			name:         "case 7: host changed but object is being deleted",
-			oldCluster:   builder.BuildClusterAsJson(builder.Name("ab123"), builder.WithDeletionTimestamp(), builder.ControlPlaneEndpoint("api.ab123.test.westeurope.azure.gigantic.io", 443)),
-			newCluster:   builder.BuildClusterAsJson(builder.Name("ab123"), builder.WithDeletionTimestamp(), builder.ControlPlaneEndpoint("api.azure.gigantic.io", 443)),
+			oldCluster:   builder.BuildCluster(builder.Name("ab123"), builder.WithDeletionTimestamp(), builder.ControlPlaneEndpoint("api.ab123.test.westeurope.azure.gigantic.io", 443)),
+			newCluster:   builder.BuildCluster(builder.Name("ab123"), builder.WithDeletionTimestamp(), builder.ControlPlaneEndpoint("api.azure.gigantic.io", 443)),
 			errorMatcher: nil,
 		},
 	}
@@ -134,12 +132,22 @@ func TestClusterUpdateValidate(t *testing.T) {
 				}
 			}
 
-			admit := &UpdateValidator{
-				logger: newLogger,
+			fakeK8sClient := unittest.FakeK8sClient()
+			ctrlClient := fakeK8sClient.CtrlClient()
+
+			handler, err := NewWebhookHandler(WebhookHandlerConfig{
+				BaseDomain: "k8s.test.westeurope.azure.gigantic.io",
+				CtrlClient: ctrlClient,
+				CtrlReader: ctrlClient,
+				Decoder:    unittest.NewFakeDecoder(),
+				Logger:     newLogger,
+			})
+			if err != nil {
+				t.Fatal(err)
 			}
 
-			// Run admission request to validate AzureConfig updates.
-			err = admit.Validate(context.Background(), getUpdateAdmissionRequest(tc.oldCluster, tc.newCluster))
+			// Run validating webhook handler on Cluster update.
+			err = handler.OnUpdateValidate(context.Background(), tc.oldCluster, tc.newCluster)
 
 			// Check if the error is the expected one.
 			switch {
@@ -154,24 +162,4 @@ func TestClusterUpdateValidate(t *testing.T) {
 			}
 		})
 	}
-}
-
-func getUpdateAdmissionRequest(oldCR []byte, newCR []byte) *v1beta1.AdmissionRequest {
-	req := &v1beta1.AdmissionRequest{
-		Resource: metav1.GroupVersionResource{
-			Version:  "cluster.x-k8s.io/v1alpha3",
-			Resource: "cluster",
-		},
-		Operation: v1beta1.Update,
-		Object: runtime.RawExtension{
-			Raw:    newCR,
-			Object: nil,
-		},
-		OldObject: runtime.RawExtension{
-			Raw:    oldCR,
-			Object: nil,
-		},
-	}
-
-	return req
 }