+++ pkg/key/key.go
@@ -2,6 +2,11 @@ package key
 
 import (
 	"fmt"
+
+	"github.com/giantswarm/microerror"
+	capi "sigs.k8s.io/cluster-api/api/v1alpha3"
+
+	"github.com/giantswarm/azure-admission-controller/internal/errors"
 )
 
 const (
@@ -39,4 +44,17 @@ func OSDiskCachingType() string {
 
 func MasterSubnetName(clusterName string) string {
 	return fmt.Sprintf("%s-%s-%s", clusterName, "VirtualNetwork", "MasterSubnet")
+}
+
+func ToClusterPtr(v interface{}) (*capi.Cluster, error) {
+	if v == nil {
+		return nil, microerror.Maskf(errors.WrongTypeError, "expected '%T', got '%T'", &capi.Cluster{}, v)
+	}
+
+	customObjectPointer, ok := v.(*capi.Cluster)
+	if !ok {
+		return nil, microerror.Maskf(errors.WrongTypeError, "expected '%T', got '%T'", &capi.Cluster{}, v)
+	}
+
+	return customObjectPointer, nil
 }